Proxmox VE highlight (#3)
* Proxmox VE highlight

Proxmox provider for Terraform by Telmate

* Execute Terraform Plan to be easy in Geany

$ terraform plan 
there is no interactive action, so may be happy to see it in the context ot the Build menu

* Build menu entries - Plan, Destroy, Apply

While Apply and Destroy must be interactive - the are executed in xterm. Geany has limit of two execute commands (entries).
Added -no-color for Plan, because looks good without escape characters (terminal colors) in the message area. To see the Message Windows, you must enable it in the View menu.